analyzer/requirements.txt
paho-mqtt==2.1.0
scikit-learn==1.5.2
numpy==2.1.2
pandas==2.2.3
joblib==1.4.2
flask==3.0.3
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.9.2
